Issue #14: New metadata handling.
By default the keyword arguments to a processor are used as metadata.

We did some quick tests these test shows that creating a json string and pickling it is considerably faster than pickling the metadata as python objects. As there are very few use cases where we use the data we optimized for speed here. We also added hashes they are not used now, but they allow for intelligent handling of the metadata.