Issue #21: Evaluation of results after previous fix showed errors with final chunk alignment and the need to mark patches as finalized or unfinalized.
Patchprocessor no offers an extra export format 'markedpatches' which contains inofrmation neccesary to distinguish between finalized and unfinalized patches. Consistency in working with patches makes it desirable to reserve the 0 serial_number for technical purposes, and therefore not use it for a patch. This change forced changes in test_patch_merging.py which was based on the assumption that serial_numbers start at 0.

Work on this problem also revealed that for wav-files the timestamp are set to the end of set of frames. Made chanhe to now correctly refer to the begin of set of frames.